Hardware Instructions
Sensors
- During the main stage of the game, input from a push button toggles the cat’s position between the top row and the bottom row. The button also allows the player to click through through the start, instruction, and endscreen stages of the game.
- We used a potentiometer to control the difficulty of the game. Obstacles move slowly if the potentiometer is set to the leftmost third, faster if it is turned to the middle third, and the fastest for the rightmost third.
- Another potentiometer allows the user to control the brightness of the screen. 

Actuators
- An LCD displays the game.
- A green LED ashes when the player scores a point.
- A piezo emits a sound when the player collides with an obstacle.

Arduino pins 4-7 and 11-12 are connected to LCD pins 11-14 and 4 and 6, respectively. LCD pins 2 and 15 are connected to 5V, pin 15 through a 220 Ohm resistor. LCD pins 1, 5, and 16 are connected to ground. LCD pin 3 receives analog input from a potentiometer. Arduino pin A1 receives analog input from the potentiometer designated to control game difficulty. Pin 2 takes a digital input from a button, and pin 3 outputs into a circuit that turns on a green LED when pin 3 is HIGH. Pins 9 and 10 connect the piezo in a circuit with a 220 Ohm resistor. It was important to connect the piezo this way, because it was too quiet when the resistance was too high, and the toneAC library required that the piezo be connected to pins 9 and 10.